Support Vector Machine(SVM) Plots
Support Vector Machine(SVM) Plots will show the pictorial Representation of a predictions. These plts are similar to Logistic Regression because we choose the kernel to be Linear in the code. To play around with the code. You can use different Kernels to make better Predictions Such as rbf, poly etc.